Avoid a NPE by caching the parent column for a cell
This fixes a NullPointerException where parentColumn was null if none of
the cells in this column were focused/unfocused before the column's
variable was removed by a script call. Because the NPE was triggered
during evaluation of a script, the stack trace isn't shown to the user,
making it harder to track this down. I may try to address that in
another PR.