Fix cell font name and error message
The error text was referring only to the default font path, obscuring
the actual error, which was that unifont.ttf is actually named
unifont-10.0.07. This change appears to have always been broken, and may
not be desirable on macOS anyway because it changes the cell data font
to a less clear monospace font.